feat: 상품 옵션 저장소의 유즈케이스를 테스트하는 로직을 추가합니다. #14
